Make FastGaussQuadrature an extension package
